+++ scripts/patches/core/all-linux/main-process/remote-control/patch.js
@@ -1,0 +1,15 @@
+"use strict";
+
+const {
+  applyLinuxRemoteControlConfigPreservationPatch,
+} = require("../../../../main-process.js");
+
+module.exports = [
+  {
+    id: "linux-remote-control-config-preservation",
+    phase: "main-bundle",
+    order: 240,
+    ciPolicy: "optional",
+    apply: applyLinuxRemoteControlConfigPreservationPatch,
+  },
+];